// projects/adminportal/src/app/modules/CMS/models/cms.ts
export interface CMSList {
  titleEn: string;
  titleAr: string;
  descriptionEn?: string; 
  descriptionAr?: string;
  contentEn?: string;
  contentAr?: string;
  published: boolean;
  createdOn?: Date;
  }
  
  export interface AddCMS {
    titleEn: string;
    titleAr: string;
    descriptionEn?: string; 
    descriptionAr?: string;
    contentEn?: string;
    contentAr?: string;
  }
  